data-normalization/requirements.txt
quixstreams<2.9
prettytable
python-dotenv